requirements.txt
openai
firecrawl-py
moviepy
faster-whisper
torch
requests
opencv-python
pillow
python-dotenv